+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+\## v1.1.0
+
+* Add DLC Content Switch to QuestModSettings
+* Add DLC Gear Items/Rewards to GearLookup.json
+* Quest objective GUI now shows the friendly JSON key name ("LilysPancakes") instead of the stripped spawn name ("CookedPancakePeach")
+* Quest rewards notification show now the friendly JSON key name ("LilysPancakes") instead of the stripped spawn name ("GEAR\_CookedPancakePeach")
+* Correct wrong prefab Names in GearLookup.json
+
+
+
 \## v1.0.0
 
 \- Initial release